+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ Changelog
 * Fix major bug in :command:`run-chip-pipeline` command where ``--impute`` argument is essentially ignored.
 * :issue:`68`: Fix bug in :meth:`api.utils.estimate_phase_beagle` method when there are no overlapping variants between input VCF and reference panel.
 * Update :meth:`api.utils.estimate_phase_beagle` method to warn when statistical phasing is skipped.
+* Upgrade Beagle version from v5.2 (beagle.28Jun21.220.jar) to v5.4 (beagle.22Jul22.46e.jar) due to a bug in v5.2.
+* Update :meth:`api.utils.estimate_phase_beagle` method to filter out variants with improper allele ('I', 'D', 'N', '.'). Note that this issue is specific to chip data.
 
 0.17.0 (2022-07-12)
 -------------------